requirements.txt
Django==2.2.3
django-bootstrap4==0.0.8
djangorestframework==3.9.4
pkg-resources==0.0.0
pytz==2019.1
sqlparse==0.3.0
